Add initial ActionEffectStack implementation
The ActionEffectStack is implemented as a stack of functions that must
be resolved to complete play of an Action card. A stack consists of
three types of items:
- Source: Denotes that a specific player must make a card selection
  according to a condition.
- Sink: A functions that modifies the Arena state in reaction to a
  previous card selection of a previous Source.
- Function: A function that modifies the Arena state, generally for a
  specific player.

When resolving an ActionEffectStack, the same parameters are used to
evaluate all items until a Sink condition is not satisfied.

This commit defines the ActionEffectStack for:
- Market
- Militia

Note that a BuyEffectStack will be necessary for some Action cards (i.e.
Merchant).