Reduce public methods for Arena
Reduce number of public methods for Arena as it tracks state, allowing
methods to be context dependent. Viewing different areas of the game
(e.g. a player's hand or supply) is now done with arena::view().

Rename as_buy_phase to to_buy_phase as it returns an owned buy phase.
Add unwrapping methods for LocationView if the variants is known.