Add Location trait
This commit implements the `Location` for `CardVec` and `Supply` to
simplify moving cards. This moves the logic of moving cards out of arena
and into the `Location` trait and trait implementations. This commit also
redefines `CardVec` as a shrinkwrapped type.

A few declarative macros are defined for use in moving cards and in
creating a `CardVec` in tests.

Unit tests and the integration test are updated. The `command` mod and
display of cards has been disabled and needs to be fixed.